chore: move apiversion package into tools/foas
Branch 2 of the cli/foas split. Moves the apiversion package from tools/cli/internal/apiversion into the tools/foas library, updates all 20 cli callers to import from tools/foas/apiversion, deletes the dead tools/cli/pkg/apiversion wrapper, and adds the require + replace directives to tools/cli/go.mod.
