+++ tools/cli/internal/openapi/filter/extension_test.go
@@ -19,7 +19,7 @@ import (
 	"testing"
 
 	"github.com/getkin/kin-openapi/openapi3"
-	"github.com/mongodb/openapi/tools/cli/internal/apiversion"
+	"github.com/mongodb/openapi/tools/foas/apiversion"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
