+++ tools/cli/internal/openapi/filter/info.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 
 	"github.com/getkin/kin-openapi/openapi3"
-	"github.com/mongodb/openapi/tools/cli/internal/apiversion"
+	"github.com/mongodb/openapi/tools/foas/apiversion"
 )
 
 // InfoVersioningFilter modifies the Info object in the OpenAPI spec with the target version.
