+++ tools/cli/internal/openapi/filter/versioning.go
@@ -19,7 +19,7 @@ import (
 	"strings"
 
 	"github.com/getkin/kin-openapi/openapi3"
-	"github.com/mongodb/openapi/tools/cli/internal/apiversion"
+	"github.com/mongodb/openapi/tools/foas/apiversion"
 )
 
 // VersioningFilter is a filter that modifies the OpenAPI spec by removing paths, operations and responses
