+++ tools/cli/internal/openapi/filter/versioning_extension.go
@@ -19,7 +19,7 @@ import (
 	"time"
 
 	"github.com/getkin/kin-openapi/openapi3"
-	"github.com/mongodb/openapi/tools/cli/internal/apiversion"
+	"github.com/mongodb/openapi/tools/foas/apiversion"
 )
 
 // VersioningExtensionFilter is a filter that updates the x-sunset and x-xgen-version extensions to a date string
